% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rfun_matchMeanSD.R
\name{matchMeanSD}
\alias{matchMeanSD}
\title{Transform Data to Desired Mean and Standard Deviation}
\usage{
matchMeanSD(data, mean = 0, sd = 1)
}
\arguments{
\item{data}{a vector includeing data to be transformed}

\item{mean}{a value of desired mean}

\item{sd}{a value of desirred SD}
}
\value{
a vector of transformed vector
}
\description{
Transform Data to Desired Mean and Standard Deviation
}
\examples{
orig_data <- c(1,3,5,10)
trans_data <- matchMeanSD(data=orig_data, mean=100, sd=15)
print(trans_data)


}
\author{
Jiangtao Gou

Fengqing Zhang
}
